+++ src/core/prelude/i18n/interface.ts
@@ -1,0 +1,13 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+export interface Locale {
+	value: string;
+	isDefined: boolean;
+	isInitialized: Promise<void>;
+}